Fixed get category by id and delete category

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -4,11 +4,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.yearup.data.mysql.interfaces.CategoryDao;
 import org.yearup.data.mysql.interfaces.ProductDao;
 import org.yearup.models.Category;
 import org.yearup.models.Product;
 
+import java.sql.SQLException;
 import java.util.List;
 
 
@@ -21,10 +23,9 @@ public class CategoriesController
     private ProductDao productDao;
 
     @Autowired
-    public CategoriesController(CategoryDao categoryDao){
+    public CategoriesController(CategoryDao categoryDao, ProductDao productDao){
         this.categoryDao = categoryDao;
-
-
+        this.productDao = productDao;
     }
 
 
@@ -38,7 +39,24 @@ public class CategoriesController
     @PreAuthorize("permitAll()")
     @RequestMapping(path = "/{id}", method = RequestMethod.GET)
     public Category getById(@PathVariable int id) {
-        return categoryDao.getById(id);
+        {
+            Category category = null;
+            try
+            {
+                category = categoryDao.getById(id);
+            }
+            catch(Exception ex)
+            {
+                throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+            }
+
+            if(category == null)
+            {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
+
+            return category;
+        }
     }
 
     // the url to return all products in category 1 would look like this
@@ -57,7 +75,7 @@ public class CategoriesController
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @RequestMapping(method = RequestMethod.PUT)
+    @PutMapping(path = "/{id}")
     // add annotation to ensure that only an ADMIN can call this function
     public void updateCategory(@PathVariable int id, @RequestBody Category category) {
         categoryDao.update(id, category);
@@ -65,9 +83,20 @@ public class CategoriesController
 
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
-    @RequestMapping(path = "/{id}", method = RequestMethod.DELETE)
+    @DeleteMapping(path = "/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
     // add annotation to ensure that only an ADMIN can call this function
     public void deleteCategory(@PathVariable int id) {
-        categoryDao.delete(id);
+        try {
+            Category category = categoryDao.getById(id);
+
+            if (category == null) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            }
+
+            categoryDao.delete(id);
+        } catch (Exception ex) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Oops... our bad.");
+        }
     }
 }
